requirements.txt
torch
torchvision
torchaudio
numpy
tqdm
PyYAML
matplotlib
scikit-learn
pandas
pytest

tensorboard

tensorboard
